+++ after/ftplugin/markdown.lua
@@ -1,5 +1,5 @@
 -- maps
 vim.keymap.set("i", "<C-]>", "- [ ] ")
 
-local snips = require("md.luasnip")
+local snips = require("pkg.luasnip")
 snips.set_language_snippets("markdown", snips.markdown_snips)
