+++ after/ftplugin/python.lua
@@ -5,7 +5,7 @@ vim.keymap.set("i", "<C-.>", "->", opts)  -- type hint return
 -- snippets
 local luasnip = require("luasnip")
 local make_snip = luasnip.parser.parse_snippet
-local md_snips = require("md.luasnip")
+local md_snips = require("pkg.luasnip")
 
 local python_snips = {
     make_snip("ia", "import $1 as $0")
