+++ init.lua
@@ -3,23 +3,23 @@ local modules  = {
     "options",
     "maps",
     -- third-party
-    "md.alpha",
-    "md.colors",
-    "md.cmp",
-    "md.git",
-    "md.lsp",
-    "md.oil",
-    "md.pairs",
-    "md.projects",
-    "md.quarto",
-    "md.slime",
-    "md.supermaven",
-    "md.telescope",
-    "md.treesitter",
-    "md.whichkey",
-    "md.surround",
-    "md.splitjoin",
-    "md.symlink",
+    "pkg.alpha",
+    "pkg.colors",
+    "pkg.cmp",
+    "pkg.git",
+    "pkg.lsp",
+    "pkg.oil",
+    "pkg.pairs",
+    "pkg.projects",
+    "pkg.quarto",
+    "pkg.slime",
+    "pkg.supermaven",
+    "pkg.telescope",
+    "pkg.treesitter",
+    "pkg.whichkey",
+    "pkg.surround",
+    "pkg.splitjoin",
+    "pkg.symlink",
     -- term-edit?
     -- zen
     -- trouble(?)
