+++ lua/pkg/slime.lua
