+++ lua/pkg/telescope.lua
